+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## UNRELEASED
+* Added `Assert::nonNegativeInteger()` and `Assert::negativeInteger()`
 
 ## 1.11.0